+++ package.json
@@ -3,7 +3,6 @@
   "version": "1.3.1",
   "main": "lib/umd/index.umd.js",
   "module": "lib/es/index.es.js",
-  "jsnext:main": "lib/es/index.es.js",
   "sideEffects": false,
   "repository": "https://github.com/gemcook/pagination.git",
   "author": "Gemcook, Inc.",
@@ -21,24 +20,24 @@
     "react-pagination"
   ],
   "devDependencies": {
-    "@babel/cli": "7.0.0-beta.40",
-    "@babel/core": "7.0.0-beta.40",
-    "@babel/plugin-external-helpers": "7.0.0-beta.40",
-    "@babel/plugin-proposal-class-properties": "7.0.0-beta.40",
-    "@babel/plugin-proposal-decorators": "7.0.0-beta.40",
-    "@babel/plugin-proposal-export-default-from": "7.0.0-beta.40",
-    "@babel/plugin-proposal-export-namespace-from": "7.0.0-beta.40",
-    "@babel/plugin-proposal-optional-chaining": "7.0.0-beta.40",
-    "@babel/plugin-syntax-decorators": "7.0.0-beta.40",
-    "@babel/plugin-syntax-export-default-from": "7.0.0-beta.40",
-    "@babel/plugin-syntax-export-namespace-from": "7.0.0-beta.40",
-    "@babel/plugin-syntax-optional-chaining": "7.0.0-beta.40",
-    "@babel/plugin-transform-runtime": "7.0.0-beta.40",
-    "@babel/preset-env": "7.0.0-beta.40",
-    "@babel/preset-flow": "7.0.0-beta.40",
-    "@babel/preset-react": "7.0.0-beta.40",
-    "@babel/preset-stage-3": "7.0.0-beta.40",
-    "@babel/runtime": "7.0.0-beta.40",
+    "@babel/cli": "7.0.0-beta.46",
+    "@babel/core": "7.0.0-beta.46",
+    "@babel/plugin-external-helpers": "7.0.0-beta.46",
+    "@babel/plugin-proposal-class-properties": "7.0.0-beta.46",
+    "@babel/plugin-proposal-decorators": "7.0.0-beta.46",
+    "@babel/plugin-proposal-export-default-from": "7.0.0-beta.46",
+    "@babel/plugin-proposal-export-namespace-from": "7.0.0-beta.46",
+    "@babel/plugin-proposal-optional-chaining": "7.0.0-beta.46",
+    "@babel/plugin-syntax-decorators": "7.0.0-beta.46",
+    "@babel/plugin-syntax-export-default-from": "7.0.0-beta.46",
+    "@babel/plugin-syntax-export-namespace-from": "7.0.0-beta.46",
+    "@babel/plugin-syntax-optional-chaining": "7.0.0-beta.46",
+    "@babel/plugin-transform-runtime": "7.0.0-beta.46",
+    "@babel/preset-env": "7.0.0-beta.46",
+    "@babel/preset-flow": "7.0.0-beta.46",
+    "@babel/preset-react": "7.0.0-beta.46",
+    "@babel/preset-stage-3": "7.0.0-beta.46",
+    "@babel/runtime": "7.0.0-beta.46",
     "@dump247/storybook-state": "^1.4.0",
     "@storybook/addon-actions": "^3.4.0",
     "@storybook/addon-info": "3.4.0",
@@ -66,8 +65,8 @@
     "eslint-plugin-prettier": "2.6.0",
     "eslint-plugin-react": "7.7.0",
     "eslint-plugin-relay": "0.0.21",
-    "flow-bin": "0.70.0",
-    "node-sass": "4.8.3",
+    "flow-bin": "0.72.0",
+    "node-sass": "4.9.0",
     "postcss": "6.0.21",
     "postcss-flexbugs-fixes": "3.3.0",
     "postcss-loader": "2.1.3",
@@ -75,20 +74,18 @@
     "prettier-eslint": "8.8.1",
     "prettier-stylelint": "0.4.2",
     "resolve-url-loader": "2.3.0",
-    "rollup": "0.58.0",
+    "rollup": "0.58.2",
     "rollup-plugin-babel": "4.0.0-beta.4",
     "rollup-plugin-closure-compiler-js": "1.0.6",
-    "rollup-plugin-commonjs": "9.1.0",
+    "rollup-plugin-commonjs": "9.1.3",
     "rollup-plugin-hypothetical": "2.1.0",
-    "rollup-plugin-json": "2.3.0",
+    "rollup-plugin-json": "3.0.0",
     "rollup-plugin-node-resolve": "3.3.0",
-    "rollup-plugin-peer-deps-external": "2.1.0",
     "rollup-plugin-postcss": "1.6.1",
     "rollup-plugin-prettier": "0.4.0",
     "rollup-plugin-replace": "2.0.0",
-    "rollup-plugin-scss": "0.4.0",
     "rollup-plugin-strip-banner": "0.2.0",
-    "rollup-plugin-url": "1.3.0",
+    "rollup-plugin-url": "1.4.0",
     "sass-loader": "6.0.7",
     "style-loader": "0.20.3",
     "stylelint": "9.1.3",
@@ -96,8 +93,8 @@
     "url-loader": "1.0.1"
   },
   "scripts": {
-    "build:umd": "node rollup.umd.build.js",
-    "build:es": "node rollup.es.build.js",
+    "build:umd": "node ./scripts/rollup/umd.build.js",
+    "build:es": "node ./scripts/rollup/es.build.js",
     "prepublish": "make prepublish",
     "storybook": "start-storybook -p 6006",
     "build-storybook": "build-storybook"
