+++ scripts/rollup/utils/resolvePath.js
@@ -1,0 +1,9 @@
+/* @flow */
+const path = require('path');
+const fs = require('fs');
+
+const appDirectory = fs.realpathSync(process.cwd());
+
+module.exports = function resolvePath(relativePath: string) {
+  return path.resolve(appDirectory, relativePath);
+};
